Rewrite the C-vm code
I modified almost all the types and all the functions.
The code is now spread across a handful of separate files.
It has not yet been tested at all. However it does compile
without warnings.

The memory management strategy tries to simulate a functional pattern
matching of sum types: tags are used to distinguish the constructors,
unions to access the varying content, and functions like
match_value_with_pair try to deconstruct the value, returning in case
of success the right kind of content (here a Pair), and freeing the
value at the same time. In case of matching error, a status variable
is set appropriately, and no memory is freed.

The instructions now have each their own function that alters a
MachineState given in parameter according to the semantics of the
instruction.