Fix exec_zzz to avoid incoherent states
Before this commit, if an instruction terminated with a non-AllOk
status, there was a real possibility that part of the machine state
(ms->term and/or ms->stack) had been free()'d. Also, some
malloc()'ed objects could have been lost in a leak.

So, I modified the implementations of the various instructions to
make sure to reset the machine state to its previous values in case
of error -- also plugging the memory leaks at the same time.