ccam: Implement Garbage Collector for ValueT
As mentioned, it'd done via reference counting. It's obviously not
perfect or optimized. But it avoids the need to actually copy
recursively ValueT values: we just go through the structure
incrementing the reference counts (which I called copy_count's) then
return the input as if we'd actually made a hard copy of it.