Fix the file extensions madness
comp used to automatically add the .c to the end of the C file;
ccam/Makefile used to automatically add it too, and name the output
file with a .out no matter what.
All this automation was more uselessly complicating than helpful, so
I'm done with it: now the user specifies full filenames every time,
and also tells ccam/Makefile the name of the output file that it
wants.