Alter RmDefs so it takes an int as parameter
It's a constructor of the type instr, in encoder.ml; it is used to
implement the let-rec constructs, along with Call and AddDefs.

Now with this change, the encoder and simulator match seamlessly the
model that was seen in class -- no differences means I don't need to
explain those differences in the README anymore.