Simplify pp_value in simulator.ml
It's the function that takes the final value computed by the
simulation and prints it in a nice, Caml-like style; `pp` stands for
pretty-print by the way.

It used to additionally display the type of the value, but since the
CCAM doesn't do it, and since it's mostly superfluous, I chose to
simplify the code instead.