Add syntactic sugar `let f x y = ...` to parser
Which gets compiled into `let f = fun x -> fun y -> ...`
The same syntax for `let rec` is also allowed.

I altered the test-programs/ to make use of that new syntax, and took
the opportunity to get rid of derelict files in that folder.

I also continued writing the README.md, and I changed the format of
the messages printed by generate-all.sh to be more easily readable.