Replace List.assoc_opt by home-made version
(In `simulator.ml`). The reason is that this function is apparently a
recent addition to the standard ocaml library, and in particular, the
computers at the university library don't have it, making it
impossible to build `simu` there. So, to make this project extra-safe
I decided to whip up my own version of that function to replace it. I
put it in `tools.md`.