Replace `read` in ccam/Makefile by script
The bash command `read` seems to vary in specs from machine to
machine. Some allow you to use it without parameters, some don't and
force you to give the name of a variable into which to write what the
command reads from standard input. Some have options to display a
prompting message before reading user input, some don't.

So, to make things as platform-independent as possible, I decided to
write a tiny bash script `ccam/pause.sh` that only does two things:
it echoes a prompting message and then uses the command `read` in the
most universal fashion I could find.

All this just to pause the execution of the `make` command in between
compiling and running the built executables...