Rewrite simulator.ml to look similar to the CCAM
I split the `exec` function into `execute_next_instruction` and
`run_machine`, renamed `initial_cfg` into `blank_state`, and
introduced the types `machine_state` and `status`, all in the aim of
making the OCaml simulator close in code structure to how I wrote the
CCAM in C: naming functions and types the same way across both
implementations of the CAM as much as possible.

It'll be especially useful for when writing the README.md, to draw
parallels between both codes.