Add built-in operator `is_empty` to OCaml code
As in, a built-in operator (like fst or head) that returns true if a
value is equal to [], and returns false otherwise (even if the value
is not a list to begin with: no typechecking).
Here by OCaml code I mean the lexer, parser, encoder, simulator,
flattener and finally the codeGenerator.

I added a new file in test-programs/ to test it, called `map2.ml`. A
word of caution: I haven't implemented this new operator in the CCAM
yet, so no point in trying to compile map2.c with it yet.