Add demo file language.ml to present the language
"The language", as in, "Mini-ML", i.e. the small, OCaml-like language
that I've been working on writing compilers and simulators for in
this very project.

So, this new demo Mini-ML program tries to exhaustively show all the
syntactic features of the language, and its semantics. I plan on
including all of its content in the README.md file, but only at the
very last moment, since I'll probably end up editing it till the last
minute.