ccam: Simplify access to main_code from runtime.c
`main_code` is the pointer to CodeT that indicates where to begin
executing the instructions of the program generated by the compiler
ocaml/comp. The main() function of runtime.c needs to have it when
creating the first MachineStateT.

Turns out all was missing to make things easy was a keyword `extern`.

In consequence, ocaml/codeGenerator.ml got simplified too.